move to java.util.Random fix EnchantmentGenerator
